backend/requirements.txt
flask
flask-cors
waitress
pymongo
bcrypt
flask_jwt_extended
python-dotenv
flask-restx